Extracted initConfig to function
