+++ src/helpers/config.ts
@@ -1,11 +1,19 @@
 import { Json } from 'mylas';
 import * as findNodeModulesPath from 'find-node-modules';
+import * as normalizePath from 'normalize-path';
 import * as fs from 'fs';
-import { dirname, join } from 'path';
-import { IRawTSConfig, ITSConfig } from '../interfaces';
-import { Output } from '../utils';
+import { basename, dirname, isAbsolute, join, normalize, resolve } from 'path';
+import {
+  IConfig,
+  IProjectConfig,
+  IRawTSConfig,
+  ITSConfig,
+  ReplaceTscAliasPathsOptions
+} from '../interfaces';
+import { Output, PathCache, TrieNode } from '../utils';
+import { importReplacers } from '.';
 
-export const loadConfig = (file: string, output: Output): ITSConfig => {
+const loadConfig = (file: string, output: Output): ITSConfig => {
   if (!fs.existsSync(file)) {
     output.error(`File ${file} not found`);
     process.exit();
@@ -74,4 +82,71 @@ export function resolveTsConfigExtendsPath(ext: string, file: string): string {
       }
     }
   }
+}
+
+export async function initConfig(options: ReplaceTscAliasPathsOptions) {
+  const output = options.output ?? new Output(options.verbose);
+
+  const configFile = !options.configFile
+    ? resolve(process.cwd(), 'tsconfig.json')
+    : !isAbsolute(options.configFile)
+    ? resolve(process.cwd(), options.configFile)
+    : options.configFile;
+
+  output.assert(
+    fs.existsSync(configFile),
+    `Invalid file path => ${configFile}`
+  );
+
+  const {
+    baseUrl = './',
+    outDir,
+    declarationDir,
+    paths,
+    replacers,
+    resolveFullPaths,
+    verbose
+  } = loadConfig(configFile, output);
+
+  output.setVerbose(verbose);
+
+  if (options.resolveFullPaths || resolveFullPaths) {
+    options.resolveFullPaths = true;
+  }
+
+  const _outDir = options.outDir ?? outDir;
+  if (declarationDir && _outDir !== declarationDir) {
+    options.declarationDir ??= declarationDir;
+  }
+
+  output.assert(_outDir, 'compilerOptions.outDir is not set');
+
+  const configDir: string = normalizePath(dirname(configFile));
+
+  // config with project details and paths
+  const projectConfig: IProjectConfig = {
+    configFile: configFile,
+    baseUrl: baseUrl,
+    outDir: _outDir,
+    configDir: configDir,
+    outPath: normalizePath(normalize(configDir + '/' + _outDir)),
+    confDirParentFolderName: basename(configDir),
+    hasExtraModule: false,
+    configDirInOutPath: null,
+    relConfDirPathInOutPath: null,
+    pathCache: new PathCache(!options.watch),
+    output: output
+  };
+
+  const config: IConfig = {
+    ...projectConfig,
+    aliasTrie:
+      options.aliasTrie ?? TrieNode.buildAliasTrie(projectConfig, paths),
+    replacers: []
+  };
+
+  // Import replacers.
+  await importReplacers(config, replacers, options.replacers);
+
+  return config;
 }
