+++ src/index.ts
@@ -1,14 +1,6 @@
 import { watch } from 'chokidar';
-import { existsSync } from 'fs';
 import { sync } from 'globby';
-import * as normalizePath from 'normalize-path';
-import { basename, dirname, isAbsolute, normalize, resolve } from 'path';
-import {
-  importReplacers,
-  loadConfig,
-  replaceAlias,
-  replaceFile
-} from './helpers';
+import { initConfig, replaceAlias, replaceFile } from './helpers';
 import {
   ReplaceTscAliasPathsOptions,
   IConfig,
@@ -16,7 +8,6 @@ import {
   IProjectConfig,
   AliasReplacerArguments
 } from './interfaces';
-import { Output, PathCache, TrieNode } from './utils';
 
 // export interfaces for api use.
 export {
@@ -36,65 +27,7 @@ export async function replaceTscAliasPaths(
     aliasTrie: undefined
   }
 ) {
-  const output = options.output ?? new Output(options.verbose);
-
-  const configFile = !options.configFile
-    ? resolve(process.cwd(), 'tsconfig.json')
-    : !isAbsolute(options.configFile)
-    ? resolve(process.cwd(), options.configFile)
-    : options.configFile;
-
-  output.assert(existsSync(configFile), `Invalid file path => ${configFile}`);
-
-  const {
-    baseUrl = './',
-    outDir,
-    declarationDir,
-    paths,
-    replacers,
-    resolveFullPaths,
-    verbose
-  } = loadConfig(configFile, output);
-
-  output.setVerbose(verbose);
-
-  if (options.resolveFullPaths || resolveFullPaths) {
-    options.resolveFullPaths = true;
-  }
-
-  const _outDir = options.outDir ?? outDir;
-  if (declarationDir && _outDir !== declarationDir) {
-    options.declarationDir ??= declarationDir;
-  }
-
-  output.assert(_outDir, 'compilerOptions.outDir is not set');
-
-  const configDir: string = normalizePath(dirname(configFile));
-
-  // config with project details and paths
-  const projectConfig: IProjectConfig = {
-    configFile: configFile,
-    baseUrl: baseUrl,
-    outDir: _outDir,
-    configDir: configDir,
-    outPath: normalizePath(normalize(configDir + '/' + _outDir)),
-    confDirParentFolderName: basename(configDir),
-    hasExtraModule: false,
-    configDirInOutPath: null,
-    relConfDirPathInOutPath: null,
-    pathCache: new PathCache(!options.watch)
-  };
-
-  const config: IConfig = {
-    ...projectConfig,
-    output: output,
-    aliasTrie:
-      options.aliasTrie ?? TrieNode.buildAliasTrie(projectConfig, paths),
-    replacers: []
-  };
-
-  // Import replacers.
-  await importReplacers(config, replacers, options.replacers);
+  const config: IConfig = await initConfig(options);
 
   // Finding files and changing alias paths
   const globPattern = [
@@ -115,10 +48,10 @@ export async function replaceTscAliasPaths(
   // Count all changed files
   const replaceCount = replaceList.filter(Boolean).length;
 
-  output.info(`${replaceCount} files were affected!`);
+  config.output.info(`${replaceCount} files were affected!`);
   if (options.watch) {
-    output.setVerbose(true);
-    output.info('[Watching for file changes...]');
+    config.output.setVerbose(true);
+    config.output.info('[Watching for file changes...]');
     const filesWatcher = watch(globPattern);
     const tsconfigWatcher = watch(config.configFile);
     const onFileChange = async (file: string) =>
@@ -126,7 +59,7 @@ export async function replaceTscAliasPaths(
     filesWatcher.on('add', onFileChange);
     filesWatcher.on('change', onFileChange);
     tsconfigWatcher.on('change', () => {
-      output.clear();
+      config.output.clear();
       filesWatcher.close();
       tsconfigWatcher.close();
       replaceTscAliasPaths(options);
